+++ SpeedyWeather/src/time_stepping/steppers/general.jl
@@ -7,14 +7,14 @@ default_time_step(L::AbstractTimeStepper) = L.Δt
 """$(TYPEDSIGNATURES)
 Computes the time step in [ms]. `Δt_at_T31` is always scaled with the resolution `trunc` 
 of the model. In case `adjust_Δt_with_output` is true, the `Δt_at_T31` is additionally 
-adjusted to the closest divisor of `output_dt` so that the output time axis is keeping
-`output_dt` exactly."""
+adjusted to the closest divisor of `interval` so that the output time axis is keeping
+`interval` exactly."""
 function get_Δt_millisec(
         Δt_at_T31::TimePeriod,
         trunc,
         radius,
         adjust_with_output::Bool,
-        output_dt::TimePeriod = DEFAULT_OUTPUT_DT,
+        interval::TimePeriod = DEFAULT_OUTPUT_INTERVAL,
     )
     # linearly scale Δt with trunc+1 (which are often powers of two)
     resolution_factor = (DEFAULT_TRUNC + 1) / (trunc + 1)
@@ -25,13 +25,13 @@ function get_Δt_millisec(
     # maybe rename to _at_trunc_and_radius?
     Δt_at_trunc = Second(Δt_at_T31).value * resolution_factor * radius_factor
 
-    if adjust_with_output && (output_dt > Millisecond(0))
-        k = round(Int, Second(output_dt).value / Δt_at_trunc)
-        divisors = Primes.divisors(Millisecond(output_dt).value)
+    if adjust_with_output && (interval > Millisecond(0))
+        k = round(Int, Second(interval).value / Δt_at_trunc)
+        divisors = Primes.divisors(Millisecond(interval).value)
         sort!(divisors)
         i = findfirst(x -> x >= k, divisors)
         k_new = isnothing(i) ? k : divisors[i]
-        Δt_millisec = Millisecond(round(Int, Millisecond(output_dt).value / k_new))
+        Δt_millisec = Millisecond(round(Int, Millisecond(interval).value / k_new))
 
         # provide info when time step is significantly shortened or lengthened
         Δt_millisec_unadjusted = round(Int, 1000 * Δt_at_trunc)
@@ -75,19 +75,19 @@ set!(L::AbstractTimeStepper; Δt::Period) = set!(L, Δt)
 function calculate_Δt!(L::AbstractTimeStepper, model::AbstractModel)
     (; trunc) = model.spectral_grid
     (; radius) = model.planet
-    output_dt = get_output_dt(model.output)
+    interval = get_interval(model.output)
 
     # take radius from planet and recalculate time step and possibly adjust with output dt
-    L.Δt_millisec = get_Δt_millisec(L.Δt_at_T31, trunc, radius, L.adjust_with_output, output_dt)
+    L.Δt_millisec = get_Δt_millisec(L.Δt_at_T31, trunc, radius, L.adjust_with_output, interval)
     L.Δt_sec = L.Δt_millisec.value / 1000
     L.Δt = L.Δt_sec / radius
 
     # check how time steps from time integration and output align
-    n = round(Int, Millisecond(output_dt).value / L.Δt_millisec.value)
+    n = round(Int, Millisecond(interval).value / L.Δt_millisec.value)
     nΔt = n * L.Δt_millisec
-    if nΔt != output_dt
+    if nΔt != interval
         @warn "$n steps of Δt = $(L.Δt_millisec.value)ms yield output every " *
-            "$(nΔt.value)ms (=$(nΔt.value / 1000)s), but output_dt = $(output_dt.value)s"
+            "$(nΔt.value)ms (=$(nΔt.value / 1000)s), but interval = $(interval.value)s"
     end
     return nothing
 end